Moved subject select functionality to separate file
