+++ assets/js/index.js
@@ -1,9 +1,11 @@
 import * as govukFrontend from 'govuk-frontend'
 import * as mojFrontend from '@ministryofjustice/frontend'
 import MapComponent from './map'
+import initSubjectSelect from './subjectSelect'
 
 govukFrontend.initAll()
 mojFrontend.initAll()
+initSubjectSelect()
 
 const $maps = document.querySelectorAll('[data-module="app-map"]')
 
@@ -20,16 +22,4 @@ function nodeListForEach(nodes, callback) {
 
 nodeListForEach($maps, $map => {
   new MapComponent($map).init()
-})
-
-function updateSubjectSelectInputs(radio) {
-  document.getElementById('subjectOrderStartDate').value = radio.dataset.start
-  document.getElementById('subjectOrderEndDate').value = radio.dataset.end
-  document.getElementById('continue').disabled = false
-}
-
-document.addEventListener('DOMContentLoaded', () => {
-  document.querySelectorAll('input[type="radio"][name="personId"]').forEach(radio => {
-    radio.addEventListener('change', () => updateSubjectSelectInputs(radio))
-  })
 })
